# TPC-H: Don't sort binary tables again, use cache by default