--- a/scripts/test/hyriseBenchmarkFileBased_test.py
+++ b/scripts/test/hyriseBenchmarkFileBased_test.py
@@ -89,7 +89,6 @@ def main():
   benchmark.expect_exact("Max duration per item is 10 seconds")
   benchmark.expect_exact("Warmup duration per item is 5 seconds")
   benchmark.expect_exact("Automatically verifying results with SQLite. This will make the performance numbers invalid.")
-  benchmark.expect_exact("Not caching tables as binary files")
   benchmark.expect_exact("Benchmarking queries from resources/test_data/queries/file_based/")
   benchmark.expect_exact("Running on tables from resources/test_data/tbl/file_based/")
   benchmark.expect_exact("Running subset of queries: select_statement")