--- a/scripts/test/hyriseBenchmarkJoinOrder_test.py
+++ b/scripts/test/hyriseBenchmarkJoinOrder_test.py
@@ -91,7 +91,6 @@ def main():
   benchmark.expect_exact("Max duration per item is 10 seconds")
   benchmark.expect_exact("Warmup duration per item is 2 seconds")
   benchmark.expect_exact("Automatically verifying results with SQLite. This will make the performance numbers invalid.")
-  benchmark.expect_exact("Not caching tables as binary files")
   benchmark.expect_exact("Benchmarking queries from third_party/join-order-benchmark")
   benchmark.expect_exact("Running on tables from resources/test_data/imdb_sample/")
   benchmark.expect_exact("Multi-threaded Topology:")