--- a/src/benchmarklib/benchmark_config.hpp
+++ b/src/benchmarklib/benchmark_config.hpp
@@ -41,7 +41,7 @@ class BenchmarkConfig {
   uint32_t clients = 1;
   bool enable_visualization = false;
   bool verify = false;
-  bool cache_binary_tables = true;
+  bool cache_binary_tables = false;  // Defaults to false for internal use, but the CLI sets it to true by default
   bool sql_metrics = false;
 
   static const char* description;