The binary dumps have no information if the entire table is sorted. As such, we currently re-sort and re-encode TPC-H tables. By checking if the table is already sorted, we can avoid this step and reduce the benchmark startup time from 0m30.447s to 0m16.147s. This is not the perfect solution and storing the flag in the table would be nicer, but at least this fixes #1864.

Also, this changes the default to cached binary tables, use `--dont_cache_binary_tables` to disable.

Before:
```
- Loading/Generating tables
[...]
-  Loading table "orders" from cached binary "tpch_cached_tables/sf-1.000000/orders.bin" (183 ms 491 µs)
[...]
-  Loading table "lineitem" from cached binary "tpch_cached_tables/sf-1.000000/lineitem.bin" (620 ms 853 µs)
- Loading/Generating tables done (1 s 45 ms)
- Sorting tables
-  Sorting 'lineitem' by 'l_shipdate' (8 s 452 ms)
-  Sorting 'orders' by 'o_orderdate' (1 s 515 ms)
- Sorting tables done (10 s 90 ms)
- Encoding tables if necessary
-  Encoding 'lineitem' - encoding applied (2 s 74 ms)
[...]
-  Encoding 'orders' - encoding applied (456 ms 582 µs)
[...]
- Encoding tables done (2 s 606 ms)
- Writing tables into binary files if necessary
- Writing 'lineitem' into binary file ".bin" (1 s 674 ms)
- Writing 'orders' into binary file ".bin" (1 s 201 ms)
- Writing tables into binary files done (2 s 875 ms)
```

New:
```
- Loading/Generating tables
[...]
-  Loading table "orders" from cached binary "tpch_cached_tables/sf-1.000000/orders.bin" (184 ms 333 µs)
[...]
-  Loading table "lineitem" from cached binary "tpch_cached_tables/sf-1.000000/lineitem.bin" (623 ms 31 µs)
- Loading/Generating tables done (1 s 41 ms)
- Sorting tables
-  Table 'lineitem' is already sorted by 'l_shipdate'
-  Table 'orders' is already sorted by 'o_orderdate'
- Sorting tables done (564 ms 646 µs)
- Encoding tables if necessary
-  Encoding 'lineitem' - no encoding necessary (535 ms 617 µs)
[...]
-  Encoding 'orders' - no encoding necessary (148 ms 3 µs)
[...]
- Encoding tables done (718 ms 586 µs)
- Writing tables into binary files if necessary
- Writing tables into binary files done (6 µs 168 ns)
```